test(core): Tests matrix serialisation
